mail template 관련 로직 service에서 추출
